+++ src/ezdxf/addons/drawing/properties.py
@@ -38,30 +38,34 @@ SHX_FONTS = {
     # Shape file structure is not documented, therefore replace this fonts by
     # true type fonts.
     # `None` is for: use the default font.
-    'AMGDT': None,  # Tolerance symbols
-    'AMGDT.SHX': None,
-    'COMPLEX': None,
-    'COMPLEX.SHX': None,
-    'ISOCP': None,
-    'ISOCP.SHX': None,
-    'ITALIC': None,
-    'ITALIC.SHX': None,
-    'GOTHICG': None,
-    'GOTHICG.SHX': None,
-    'GREEKC': None,
-    'GREEKC.SHX': None,
-    'ROMANS': None,
-    'ROMANS.SHX': None,
-    'SCRIPTS': None,
-    'SCRIPTS.SHX': None,
-    'SCRIPTC': None,
-    'SCRIPTC.SHX': None,
-    'SIMPLEX': None,
-    'SIMPLEX.SHX': None,
-    'SYMATH': None,
-    'SYMATH.SHX': None,
-    'TXT': None,  # Default AutoCAD font
-    'TXT.SHX': None,
+    'AMGDT': "amgdt___.ttf",  # Tolerance symbols
+    'AMGDT.SHX': "amgdt___.ttf",
+    'COMPLEX': "complex_.ttf",
+    'COMPLEX.SHX': "complex_.ttf",
+    'ISOCP': "isocp.ttf",
+    'ISOCP.SHX': "isocp.ttf",
+    'ITALIC': "italicc_.ttf",
+    'ITALIC.SHX': "italicc_.ttf",
+    'GOTHICG': "gothicg_.ttf",
+    'GOTHICG.SHX': "gothicg_.ttf",
+    'GREEKC': "greekc.ttf",
+    'GREEKC.SHX': "greekc.ttf",
+    'ROMANS': "romans__.ttf",
+    'ROMANS.SHX': "romans__.ttf",
+    'SCRIPTS': "scripts_.ttf",
+    'SCRIPTS.SHX': "scripts_.ttf",
+    'SCRIPTC': "scriptc_.ttf",
+    'SCRIPTC.SHX': "scriptc_.ttf",
+    'SIMPLEX': "simplex_.ttf",
+    'SIMPLEX.SHX': "simplex_.ttf",
+    'SYMATH': "symath__.ttf",
+    'SYMATH.SHX': "symath__.ttf",
+    'SYMAP': "symap___.ttf",
+    'SYMAP.SHX': "symap___.ttf",
+    'SYMETEO': "symeteo_.ttf",
+    'SYMETEO.SHX': "symeteo_.ttf",
+    'TXT': "monotxt_.ttf",  # Default AutoCAD font
+    'TXT.SHX': "monotxt_.ttf",
 }