+++ NEWS.md
@@ -9,7 +9,7 @@ Version 0.15a3 - dev
 - NEW: linetype support for matplotlib drawing backend
 - NEW: HATCH island and basic pattern support for matplotlib drawing backend
 - NEW: HATCH island support for pyqt drawing backend
-- NEW: Font support for matplotlib drawing backend
+- NEW: Font support for matplotlib- and pyqt drawing backend
 - NEW: recover misplaced tags of the `AcDbEntity` subclass (color, layer, 
   linetype, ...), supported by all loading modes
 - NEW: `ezdxf.addons.geo` module, support for the